ci: add rpcimportable test (#2817)
* ci: add rpcimportable test

* add ci

* fmt

* use github.com/btcsuite/btcd/btcutil in pkg/chains

* remove app imports types tests

* use standalone sdkconfig package

* fix policies test

* move crosschain keeper tests from types to keeper

* never seal config in tests

* use https://github.com/zeta-chain/ethermint/pull/126

* add some comments

* use merged ethermint hash

* show resulting go.mod
