+++ go.mod
@@ -57,7 +57,7 @@ require (
 	github.com/spf13/pflag v1.0.5
 	github.com/spf13/viper v1.16.0
 	github.com/stretchr/testify v1.9.0
-	github.com/zeta-chain/ethermint v0.0.0-20240909234716-2fad916e7179
+	github.com/zeta-chain/ethermint v0.0.0-20240927155358-f34e2a4a86f1
 	github.com/zeta-chain/keystone/keys v0.0.0-20240826165841-3874f358c138
 	github.com/zeta-chain/protocol-contracts v1.0.2-athens3.0.20240924201108-3a274ce7bad0
 	gitlab.com/thorchain/tss/go-tss v1.6.5
