+++ pkg/rpc/clients_crosschain.go
@@ -2,15 +2,12 @@ package rpc
 
 import (
 	"context"
-	"sort"
 
 	"cosmossdk.io/errors"
-	"github.com/cosmos/cosmos-sdk/types/query"
 	"google.golang.org/grpc"
 
 	"github.com/zeta-chain/node/pkg/chains"
 	"github.com/zeta-chain/node/x/crosschain/types"
-	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
 )
 
 // 32MB
@@ -156,39 +153,4 @@ func (c *Clients) GetInboundTrackersForChain(ctx context.Context, chainID int64)
 	}
 
 	return resp.InboundTracker, nil
-}
-
-// GetAllOutboundTrackerByChain returns all outbound trackers for a chain
-func (c *Clients) GetAllOutboundTrackerByChain(
-	ctx context.Context,
-	chainID int64,
-	order interfaces.Order,
-) ([]types.OutboundTracker, error) {
-	in := &types.QueryAllOutboundTrackerByChainRequest{
-		Chain: chainID,
-		Pagination: &query.PageRequest{
-			Key:        nil,
-			Offset:     0,
-			Limit:      2000,
-			CountTotal: false,
-			Reverse:    false,
-		},
-	}
-
-	resp, err := c.Crosschain.OutboundTrackerAllByChain(ctx, in)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get all outbound trackers")
-	}
-
-	if order == interfaces.Ascending {
-		sort.SliceStable(resp.OutboundTracker, func(i, j int) bool {
-			return resp.OutboundTracker[i].Nonce < resp.OutboundTracker[j].Nonce
-		})
-	} else if order == interfaces.Descending {
-		sort.SliceStable(resp.OutboundTracker, func(i, j int) bool {
-			return resp.OutboundTracker[i].Nonce > resp.OutboundTracker[j].Nonce
-		})
-	}
-
-	return resp.OutboundTracker, nil
 }
