+++ x/authority/types/policies_test.go
@@ -3,28 +3,15 @@ package types_test
 import (
 	"testing"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 
-	"github.com/zeta-chain/node/app"
+	"github.com/zeta-chain/node/pkg/sdkconfig"
 	"github.com/zeta-chain/node/testutil/sample"
 	"github.com/zeta-chain/node/x/authority/types"
 )
 
-// setConfig sets the global config to use zeta chain's bech32 prefixes
-func setConfig(t *testing.T) {
-	defer func(t *testing.T) {
-		if r := recover(); r != nil {
-			t.Log("config is already sealed", r)
-		}
-	}(t)
-	cfg := sdk.GetConfig()
-	cfg.SetBech32PrefixForAccount(app.Bech32PrefixAccAddr, app.Bech32PrefixAccPub)
-	cfg.Seal()
-}
-
 func TestPolicies_Validate(t *testing.T) {
-	setConfig(t)
+	sdkconfig.SetDefault(false)
 	// use table driven tests to test the validation of policies
 	tests := []struct {
 		name        string
